+++ src/zombiebehavior.h
@@ -17,7 +17,7 @@ namespace zombie {
 		ZombieBehavior(Unit* unit);
 		~ZombieBehavior();
 
-		void updateInput(float time, float deltaTime) override;
+		void updateInput(double time, double deltaTime) override;
 
 		MovingObject* getMovingObject() const;
